Cleanup the readme and remove some cluttering
Prior to this change, the readme wasn't available and also
the requirements were missing

This change add a basic readme, add a requirements.txt and remove some
files that are not useful or part of the game
