+++ README.md
@@ -1,0 +1,10 @@
+# GhostRunner
+
+My First game!
+
+## First Start
+
+```
+pip -r requirements.txt
+python ./ghostrunner.py
+```
