+++ requirements.txt
@@ -1,0 +1,1 @@
+pygame==2.5.0
